Removed set of destination for ACK packets in TOSSIM rfxlink
This was done in old TOSSIM rfxlink code which relied on
TossimPacketModelC to deliver the packet. Now TossimDriverLayerP takes
care of everything from the CpmModelC and hence the destination does
not need to be set for a packet to arrive.

Setting the destination was kinda of hack and its good that we can
remove this.